+++ src/test/java/teammates/test/cases/ui/browsertests/InstructorFeedbackEditPageUiTest.java
@@ -1,11 +1,13 @@
 package teammates.test.cases.ui.browsertests;
 
 import static org.testng.AssertJUnit.assertEquals;
+import static org.testng.AssertJUnit.assertFalse;
 import static org.testng.AssertJUnit.assertNotNull;
 import static org.testng.AssertJUnit.assertNull;
 import static org.testng.AssertJUnit.assertTrue;
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.WebElement;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
@@ -13,6 +15,7 @@ import org.testng.annotations.Test;
 import teammates.common.datatransfer.DataBundle;
 import teammates.common.datatransfer.FeedbackSessionAttributes;
 import teammates.common.util.Const;
+import teammates.common.util.ThreadHelper;
 import teammates.common.util.TimeHelper;
 import teammates.common.util.Url;
 import teammates.test.driver.AssertHelper;
@@ -20,6 +23,7 @@ import teammates.test.driver.BackDoor;
 import teammates.test.driver.TestProperties;
 import teammates.test.pageobjects.Browser;
 import teammates.test.pageobjects.BrowserPool;
+import teammates.test.pageobjects.FeedbackQuestionSubmitPage;
 import teammates.test.pageobjects.FeedbackSubmitPage;
 import teammates.test.pageobjects.InstructorFeedbackEditPage;
 import teammates.test.pageobjects.InstructorFeedbacksPage;
@@ -33,7 +37,7 @@ import com.google.appengine.api.datastore.Text;
  * SUT is {@link InstructorFeedbackEditPage}.
  */
 @Priority(-1)
-public class InstructorFeedbackEditSessionsUiTest extends BaseUiTestCase {
+public class InstructorFeedbackEditPageUiTest extends BaseUiTestCase {
     private static Browser browser;
     private static InstructorFeedbackEditPage feedbackEditPage;
     private static DataBundle testData;
@@ -46,7 +50,7 @@ public class InstructorFeedbackEditSessionsUiTest extends BaseUiTestCase {
     @BeforeClass
     public static void classSetup() throws Exception {
         printTestClassHeader();
-        testData = loadDataBundle("/InstructorFeedbackEditSessionsUiTest.json");
+        testData = loadDataBundle("/InstructorFeedbackEditPageUiTest.json");
         removeAndRestoreTestDataOnServer(testData);
 
         editedSession = testData.feedbackSessions.get("openSession");
@@ -75,12 +79,33 @@ public class InstructorFeedbackEditSessionsUiTest extends BaseUiTestCase {
 
         testEditSessionLink();
         testEditSessionAction();
+        
+        testGeneralQuestionOperations();
 
         testPreviewSessionAction();
 
         testDoneEditingLink();
         
         testDeleteSessionAction();
+    }
+    
+    private void testGeneralQuestionOperations() {
+        testNewQuestionLink();
+        testInputValidationForQuestion();
+        testAddQuestionAction();
+
+        testEditQuestionLink();
+        testEditQuestionAction();
+
+        testGetQuestionLink();
+        testCopyQuestion();
+
+        testChangeFeedbackRecipient();
+
+        testEditQuestionNumberAction();
+
+        testDeleteQuestionAction(2);
+        testDeleteQuestionAction(1);
     }
 
     private void testContent() {
@@ -135,6 +160,162 @@ public class InstructorFeedbackEditSessionsUiTest extends BaseUiTestCase {
         feedbackEditPage.clickSaveSessionButton();
     }
 
+    private void testNewQuestionLink() {
+
+        ______TS("new question (frame) link");
+        feedbackEditPage.clickNewQuestionButton();
+        assertTrue(feedbackEditPage.verifyNewEssayQuestionFormIsDisplayed());
+    }
+
+    private void testInputValidationForQuestion() {
+
+        ______TS("empty question text");
+
+        feedbackEditPage.clickAddQuestionButton();
+        assertEquals(Const.StatusMessages.FEEDBACK_QUESTION_TEXTINVALID, feedbackEditPage.getStatus());
+
+        ______TS("empty number of max respondants field");
+
+        feedbackEditPage.fillQuestionBox("filled qn");
+        feedbackEditPage.selectRecipientsToBeStudents();
+        feedbackEditPage.fillNumOfEntitiesToGiveFeedbackToBox("");
+        feedbackEditPage.clickCustomNumberOfRecipientsButton();
+        feedbackEditPage.clickAddQuestionButton();
+        assertEquals(Const.StatusMessages.FEEDBACK_QUESTION_NUMBEROFENTITIESINVALID, feedbackEditPage.getStatus());
+
+    }
+
+    private void testAddQuestionAction() {
+
+        ______TS("add question action success");
+
+        feedbackEditPage.clickMaxNumberOfRecipientsButton();
+        feedbackEditPage.clickAddQuestionButton();
+        assertEquals(Const.StatusMessages.FEEDBACK_QUESTION_ADDED, feedbackEditPage.getStatus());
+        assertNotNull(BackDoor.getFeedbackQuestion(courseId, feedbackSessionName, 1));
+        feedbackEditPage.verifyHtmlMainContent("/instructorFeedbackQuestionAddSuccess.html");
+    }
+
+    private void testEditQuestionLink() {
+
+        ______TS("edit question link");
+
+        assertEquals(true, feedbackEditPage.clickEditQuestionButton(1));
+    }
+
+    private void testEditQuestionAction() {
+
+        ______TS("edit question 1 to Team-to-Team");
+
+        feedbackEditPage.clickVisibilityOptionsForQuestion1();
+        feedbackEditPage.selectGiverTypeForQuestion1("Teams in this course");        
+        feedbackEditPage.selectRecipientTypeForQuestion1("Other teams in the course");
+        feedbackEditPage.verifyHtmlMainContent("/instructorFeedbackQuestionEditToTeamToTeam.html");
+
+        ______TS("test visibility options of question 1");
+        feedbackEditPage.clickquestionSaveForQuestion1();
+        feedbackEditPage.clickVisibilityOptionsForQuestion1();
+        //TODO: use simple element checks instead of html checks after adding names to the checkboxes 
+        //      in the edit page (follow todo in instructorsFeedbackEdit.js)
+        feedbackEditPage.verifyHtmlMainContent("/instructorFeedbackQuestionVisibilityOptions.html");
+        
+        ______TS("test visibility preview of question 1");
+        feedbackEditPage.clickVisibilityPreviewForQuestion1();
+        WebElement visibilityMessage = browser.driver.findElement(By.className("visibilityMessageButton"));
+        feedbackEditPage.waitForElementVisible(visibilityMessage);
+
+        feedbackEditPage.verifyHtmlMainContent("/instructorFeedbackQuestionVisibilityPreview.html");
+        
+        //change back
+        feedbackEditPage.clickVisibilityOptionsForQuestion1();
+        feedbackEditPage.selectGiverTypeForQuestion1("Me (Session creator)");
+        feedbackEditPage.selectRecipientTypeForQuestion1("Other students in the course");
+        feedbackEditPage.clickquestionSaveForQuestion1();
+    }
+    
+    private void testGetQuestionLink() {
+
+        ______TS("get individual question link");
+
+        feedbackEditPage.clickGetLinkButton();
+        String questionId = BackDoor.getFeedbackQuestion(courseId,
+                feedbackSessionName, 1).getId();
+        
+        String expectedUrl = TestProperties.inst().TEAMMATES_URL
+                + Const.ActionURIs.INSTRUCTOR_FEEDBACK_QUESTION_SUBMISSION_EDIT_PAGE
+                + "?courseid=" + courseId
+                + "&fsname=First+Session"
+                + "&questionid=" + questionId;
+
+        assertTrue(feedbackEditPage.isElementVisible("statusMessage"));
+        assertEquals(
+                "Link for question 1: " + expectedUrl,
+                feedbackEditPage.getStatus());
+        
+        Url url = new Url(expectedUrl).withUserId(instructorId);
+        FeedbackQuestionSubmitPage questionPage = loginAdminToPage(browser, url, FeedbackQuestionSubmitPage.class);
+        
+        assertTrue(questionPage.isCorrectPage(courseId, feedbackSessionName));
+        
+        feedbackEditPage = getFeedbackEditPage();
+    }
+
+    private void testEditQuestionNumberAction() {
+        ______TS("edit question number success");
+
+        assertEquals(true, feedbackEditPage.clickEditQuestionButton(2));
+        feedbackEditPage.selectQuestionNumber(2, 1);        
+        feedbackEditPage.clickSaveExistingQuestionButton(2);
+        assertEquals(Const.StatusMessages.FEEDBACK_QUESTION_EDITED, feedbackEditPage.getStatus());
+    }
+
+    private void testCopyQuestion() {
+        
+        ______TS("Success case: copy questions successfully");
+        
+        ThreadHelper.waitFor(1000);
+        feedbackEditPage.clickCopyButton();
+        ThreadHelper.waitFor(1000);
+        feedbackEditPage.clickCopyTableAtRow(0);
+        feedbackEditPage.clickCopySubmitButton();
+        feedbackEditPage.verifyHtmlMainContent("/instructorFeedbackCopyQuestionSuccess.html");
+    }
+    
+    private void testChangeFeedbackRecipient() {
+        ______TS("click on Edit visibility then change recipient");
+
+        feedbackEditPage = getFeedbackEditPage();
+
+        assertTrue(feedbackEditPage.verifyPreviewLabelIsActive(1));
+        assertFalse(feedbackEditPage.verifyEditLabelIsActive(1));
+        assertTrue(feedbackEditPage.verifyVisibilityMessageIsDisplayed(1));
+        assertFalse(feedbackEditPage.verifyVisibilityOptionsIsDisplayed(1));
+
+        feedbackEditPage.clickQuestionEditForQuestion1();
+        feedbackEditPage.clickEditLabel(1);
+        feedbackEditPage.selectRecipientTypeForQuestion1("Other teams in the course");
+
+        assertFalse(feedbackEditPage.verifyPreviewLabelIsActive(1));
+        assertTrue(feedbackEditPage.verifyEditLabelIsActive(1));
+        assertFalse(feedbackEditPage.verifyVisibilityMessageIsDisplayed(1));
+        assertTrue(feedbackEditPage.verifyVisibilityOptionsIsDisplayed(1));
+    }
+
+    private void testDeleteQuestionAction(int qnNumber) {
+        
+        ______TS("qn " + qnNumber + " delete then cancel");
+
+        feedbackEditPage.clickAndCancel(feedbackEditPage.getDeleteQuestionLink(qnNumber));
+        assertNotNull(BackDoor.getFeedbackQuestion(courseId, feedbackSessionName, qnNumber));
+
+        ______TS("qn " + qnNumber + " delete then accept");
+
+        feedbackEditPage.clickAndConfirm(feedbackEditPage.getDeleteQuestionLink(qnNumber));
+        assertEquals(Const.StatusMessages.FEEDBACK_QUESTION_DELETED, feedbackEditPage.getStatus());
+        assertNull(BackDoor.getFeedbackQuestion(courseId, feedbackSessionName, qnNumber));
+        
+    }
+    
     private void testPreviewSessionAction() {
 
         // add questions for previewing
